cleanup & fix st7920 simulator (add klipper support) (#705)
* fix bug
* improve st7920 simulator for klipper

+++ TFT/src/User/API/interfaceCmd.c
@@ -191,15 +191,15 @@ void sendQueueCmd(void)
               }
             Serial_Puts(SERIAL_PORT_2, "Begin file list\n");
             if (mountFS() == true && scanPrintFiles() == true){
-              for (UINT i = 0; i < infoFile.f_num; i++) {
+              for (uint16_t i = 0; i < infoFile.f_num; i++) {
                 Serial_Puts(SERIAL_PORT_2,infoFile.file[i]);
                 Serial_Puts(SERIAL_PORT_2,"\n");
-                }
-              for (UINT i = 0; i < infoFile.F_num; i++) {
+              }
+              for (uint16_t i = 0; i < infoFile.F_num; i++) {
                 Serial_Puts(SERIAL_PORT_2,"/");
                 Serial_Puts(SERIAL_PORT_2,infoFile.folder[i]);
                 Serial_Puts(SERIAL_PORT_2,"/\n");
-                }
+              }
             }
             Serial_Puts(SERIAL_PORT_2, "End file list\nok\n");
             infoCmd.count--;
